Connect appwrite with the login screen #26
