+++ lib/src/utils/snackbar.dart
@@ -1,13 +1,11 @@
-import 'package:chamting_app/main.dart';
 import 'package:flutter/material.dart';
+
+import '../app.dart';
 
 void showSnackbar({
   required String message,
   required Color color,
 }) =>
     scaffoldMessengerKey.currentState!.showSnackBar(
-      SnackBar(
-        content: Text(message),
-        backgroundColor: color,
-      ),
+      SnackBar(content: Text(message), backgroundColor: color),
     );
